backend/requirements.txt
fastapi[standard]
python-dotenv
pydantic
pydantic[email]
motor
pymongo